moved clopath archiving node to separate file
