+++ models/aeif_cbvg_2010.h
@@ -34,7 +34,7 @@
 #include <gsl/gsl_odeiv.h>
 
 // Includes from nestkernel:
-#include "archiving_node.h"
+#include "clopath_archiving_node.h"
 #include "connection.h"
 #include "event.h"
 #include "nest_types.h"
