+++ nestkernel/CMakeLists.txt
@@ -21,6 +21,7 @@ set( nestkernel_sources
     universal_data_logger_impl.h universal_data_logger.h
     recordables_map.h
     archiving_node.h archiving_node.cpp
+    clopath_archiving_node.h clopath_archiving_node.cpp
     common_synapse_properties.h common_synapse_properties.cpp
     completed_checker.h completed_checker.cpp
     sibling_container.h sibling_container.cpp
